+++ backend/internal/model/database/pitch_media.go
@@ -1,10 +1,10 @@
 package database
 
 type PitchMedia struct {
-	ID                 *int   `json:"id,omitempty"`
-	PitchID            int    `json:"pitch_id"`
+	ID                 *int64 `json:"id,omitempty"`
+	PitchID            int64  `json:"pitch_id"`
 	URL                string `json:"url"`
 	MediaType          string `json:"media_type"`
-	OrderInDescription int    `json:"order_in_description"`
+	OrderInDescription int64  `json:"order_in_description"`
 	CreatedAt          string `json:"created_at,omitempty"`
 }
